DYN-3908-OpenControls-ByStep Code Review
I updated all the contents for each step following the wiki entry Content for Dynamo First Time Use Guide.
Also updated the Height for the Run Status popup
Also moved the functions for executing the UI Automation from the Guide class to the Step class.
