+++ src/DynamoCoreWpf/UI/GuidedTour/Guide.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Windows;
-using System.Windows.Controls;
 using System.Windows.Media;
 using Newtonsoft.Json;
 
@@ -109,12 +108,6 @@ namespace Dynamo.Wpf.UI.GuidedTour
                 if (nextStep != null)
                 {
                     nextStep.Show();
-
-                    //In case the UIAutomation info is set for the Step we execute the action when the Next button is pressed
-                    if (nextStep.UIAutomation != null)
-                    {
-                        ExecuteUIAutomationStep(nextStep.UIAutomation, true);
-                    }
                 }
 
             }
@@ -123,12 +116,6 @@ namespace Dynamo.Wpf.UI.GuidedTour
             if (CurrentStep != null)
             {
                 CurrentStep.Hide();
-
-                //Disable the current action automation that is executed for the Current Step (if there is one)
-                if (CurrentStep.UIAutomation != null)
-                {
-                    ExecuteUIAutomationStep(CurrentStep.UIAutomation, false);
-                }
             }             
         }
 
@@ -147,12 +134,6 @@ namespace Dynamo.Wpf.UI.GuidedTour
                 if (prevStep != null)
                 {
                     prevStep.Show();
-
-                    //In case the UIAutomation info is set for the Step we execute the action when the Back button is pressed
-                    if (prevStep.UIAutomation != null)
-                    {
-                        ExecuteUIAutomationStep(prevStep.UIAutomation, true);
-                    }
                 }                   
             }
 
@@ -160,12 +141,6 @@ namespace Dynamo.Wpf.UI.GuidedTour
             if (CurrentStep != null)
             {
                 CurrentStep.Hide();
-
-                //Disable the current action automation that is executed for the Current Step (if there is one)
-                if (CurrentStep.UIAutomation != null)
-                {
-                    ExecuteUIAutomationStep(CurrentStep.UIAutomation, false);
-                }
             }             
         }
 
@@ -217,29 +192,6 @@ namespace Dynamo.Wpf.UI.GuidedTour
             }
 
             return foundChild;
-        }
-
-        /// <summary>
-        /// This method will execute an UIAutomation action over a specific UIElement
-        /// </summary>
-        /// <param name="uiAutomationData">UIAutomation info read from a json file</param>
-        /// <param name="enableUIAutomation">Enable/Disable the automation action for a specific UIElement</param>
-        private void ExecuteUIAutomationStep(StepUIAutomation uiAutomationData, bool enableUIAutomation)
-        {
-            switch (uiAutomationData.ControlType.ToUpper())
-            {
-                case "MENUITEM":
-                    if (uiAutomationData.UIElementAutomation != null)
-                    {
-                        if(uiAutomationData.Action.ToUpper().Equals("OPEN"))
-                        {
-                            MenuItem menuEntry = uiAutomationData.UIElementAutomation as MenuItem;
-                            menuEntry.IsSubmenuOpen = enableUIAutomation;
-                            menuEntry.StaysOpenOnClick = enableUIAutomation;
-                        }                      
-                    }                      
-                    break;
-            }
         }
     }
 }
