+++ src/DynamoCoreWpf/UI/GuidedTour/Step.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Windows.Controls;
 using System.Windows.Controls.Primitives;
 using System.Windows.Media;
 using Newtonsoft.Json;
@@ -128,6 +129,12 @@ namespace Dynamo.Wpf.UI.GuidedTour
         public void Show()
         {
             stepUIPopup.IsOpen = true;
+
+            //In case the UIAutomation info is set for the Step we execute the action when the Next button is pressed
+            if (UIAutomation != null)
+            {
+                ExecuteUIAutomationStep(UIAutomation, true);
+            }
         }
 
         /// <summary>
@@ -136,6 +143,35 @@ namespace Dynamo.Wpf.UI.GuidedTour
         public void Hide()
         {
             stepUIPopup.IsOpen = false;
+
+            //Disable the current action automation that is executed for the Current Step (if there is one)
+            if (UIAutomation != null)
+            {
+                ExecuteUIAutomationStep(UIAutomation, false);
+            }
+        }
+
+        /// <summary>
+        /// This method will execute an UIAutomation action over a specific UIElement
+        /// </summary>
+        /// <param name="uiAutomationData">UIAutomation info read from a json file</param>
+        /// <param name="enableUIAutomation">Enable/Disable the automation action for a specific UIElement</param>
+        private void ExecuteUIAutomationStep(StepUIAutomation uiAutomationData, bool enableUIAutomation)
+        {
+            switch (uiAutomationData.ControlType.ToUpper())
+            {
+                case "MENUITEM":
+                    if (uiAutomationData.UIElementAutomation != null)
+                    {
+                        if (uiAutomationData.Action.ToUpper().Equals("OPEN"))
+                        {
+                            MenuItem menuEntry = uiAutomationData.UIElementAutomation as MenuItem;
+                            menuEntry.IsSubmenuOpen = enableUIAutomation;
+                            menuEntry.StaysOpenOnClick = enableUIAutomation;
+                        }
+                    }
+                    break;
+            }
         }
 
         /// <summary>
